+++ omar-huanca/lesson-1/homework-cars/cars_app/View.java
@@ -1,5 +1,9 @@
 package cars_app;
 
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Scanner;
 
 /**
@@ -9,8 +13,8 @@ import java.util.Scanner;
  * @version 1.0.
  */
 public class View {
+    private List<Car> carsList;
     private int index;
-    private CarsManager carsManager;
     private Scanner scanInput;
     private String inputString;
 
@@ -19,9 +23,8 @@ public class View {
      */
     public View() {
         index = 0;
-        carsManager = new CarsManager();
         scanInput = new Scanner(System.in);
-
+        carsList = new ArrayList<>();
     }
 
     /**
@@ -32,7 +35,7 @@ public class View {
             System.out.println("==============================");
             System.out.println("Your Car List");
             System.out.println("==============================");
-            this.carsManager.showList();
+            this.showList();
             System.out.println("==============================");
             System.out.println("Welcome please choose a option");
             System.out.println("==============================");
@@ -70,11 +73,11 @@ public class View {
         System.out.println("Please Insert a distance:");
         distance = Double.parseDouble(scanInput.nextLine());
         System.out.println("==============================");
-        Car winner = carsManager.startPerformanceTest(distance);
+        Car winner = this.startPerformanceTest(distance);
         System.out.println("=WINNER= " + winner.getName() + " =WINNER=");
         System.out.println("DISTANCE: " + distance + " KM");
         System.out.println("==============================");
-        carsManager.showPerformanceList();
+        this.showPerformanceList();
         index = 3;
 
     }
@@ -95,15 +98,78 @@ public class View {
             }
 
             // Creating Car
-            Car newCar = this.carsManager.createNewCar(inputString);
+            Car newCar = this.createNewCar(inputString);
 
             // Adding Car to List
-            this.carsManager.addCarToList(newCar);
+            this.addCarToList(newCar);
 
             System.out.println("=New Car has been created=");
             System.out.println("Name: " + newCar.getName());
             System.out.println("Gasoline Full: " + newCar.getGasoline() + " [L]");
             System.out.println("==========================");
         }
+    }
+
+    /**
+     * This method is to show car list
+     */
+    public void showList() {
+        carsList.forEach(car -> System.out.println(car.getName() + " Gasoline: " + car.getGasoline() + " Tank size: " + car.getMaxGasoline()));
+    }
+
+    /**
+     * This method is to start performance test according distance
+     *
+     * @param distance is the distance in km to move
+     * @return Car
+     */
+    public Car startPerformanceTest(double distance) {
+        // moving distance
+        carsList.forEach(car -> car.moveADistanceWithCar(distance));
+        // comparing gasoline
+        return carsList.stream()
+                .max(Comparator.comparing(Car::getGasoline))
+                .orElseThrow(NoSuchElementException::new);
+    }
+
+    /**
+     * This method is to show performance test resume
+     */
+    public void showPerformanceList() {
+        carsList.forEach(car -> {
+            System.out.println("Car Name: " + car.getName());
+            boolean state = car.getGasoline() > 0;
+            System.out.println("Performace Test Passed?: " + state);
+            System.out.println("Gasoline per KM: " + car.getGasConsume() + " [L]");
+            if (state) {
+                System.out.println("Actual Gasoline: " + car.getGasoline() + " [L]");
+            } else {
+                System.out.println("Actual Gasoline: 0");
+            }
+            System.out.println("Tank Limit: " + car.getMaxGasoline() + " [L]");
+            System.out.println("====================================");
+        });
+    }
+
+    /**
+     * This method is to add a Contact in to list of carsList
+     *
+     * @param carName new car name
+     * @return Car
+     */
+    public Car createNewCar(String carName) {
+        Car newCar = new Car();
+        newCar.setName(carName);
+        newCar.setGasoline(newCar.getMaxGasoline());
+        return newCar;
+    }
+
+    /**
+     * This method is to add a Car in to list of carsList
+     *
+     * @param carToAdd car to add in the list
+     */
+    public void addCarToList(Car carToAdd) {
+        carsList.add(carToAdd);
     }
 }
